+++ CHANGELOG.md
@@ -9,10 +9,12 @@ All notable changes to `telegram` will be documented in this file
 * Drop support for Laravel 10.
 * Drop support for PHP 8.1.
 * Add support for Laravel 12.
+* Add `sticker` method to the `TelegramFile` to send sticker file.
 * Add `sendWhen` method to conditionally send a message.
 * Add ParseMode Enum and refactor parsing mode setting logic.
 * Add `buttonWithWebApp` method to open web app from a button.
 * Add `onError` method to handle exceptions. Based of https://github.com/laravel-notification-channels/telegram/pull/201
+* Refactor `sendFile` to support raw data sending.
 * Refactor `escapedLine` method.
 * Refactor `HasSharedLogic` trait.
 * Refactor classes to use PHP 8.2 features.